拆分Http请求基础参数到基类, 让代码结构更易读.
